Move additive read distributed to free function
