+++ core/distributed/matrix.cpp
@@ -6,6 +6,7 @@
 
 #include <ginkgo/core/base/array.hpp>
 #include <ginkgo/core/base/precision_dispatch.hpp>
+#include <ginkgo/core/distributed/assembly_helpers.hpp>
 #include <ginkgo/core/distributed/vector.hpp>
 #include <ginkgo/core/matrix/coo.hpp>
 #include <ginkgo/core/matrix/csr.hpp>
@@ -22,120 +23,11 @@ namespace matrix {
 namespace {
 
 
-GKO_REGISTER_OPERATION(count_non_owning_entries,
-                       distributed_matrix::count_non_owning_entries);
-GKO_REGISTER_OPERATION(fill_send_buffers,
-                       distributed_matrix::fill_send_buffers);
 GKO_REGISTER_OPERATION(separate_local_nonlocal,
                        distributed_matrix::separate_local_nonlocal);
 
 
 }  // namespace
-
-
-template <typename ValueType, typename LocalIndexType, typename GlobalIndexType>
-void communicate_non_owning(
-    std::shared_ptr<const Executor> exec, mpi::communicator comm,
-    const Partition<LocalIndexType, GlobalIndexType>* row_partition,
-    const Partition<LocalIndexType, GlobalIndexType>* col_partition,
-    const device_matrix_data<ValueType, GlobalIndexType>& input_data,
-    device_matrix_data<ValueType, GlobalIndexType>& all_data)
-{
-    size_type num_entries = input_data.get_num_stored_elements();
-    size_type num_parts = comm.size();
-    auto use_host_buffer = mpi::requires_host_buffer(exec, comm);
-    auto global_num_rows = row_partition->get_size();
-    auto global_num_cols = col_partition->get_size();
-    auto local_part = comm.rank();
-    dim<2> global_dim{global_num_rows, global_num_cols};
-    array<comm_index_type> send_sizes{exec, num_parts};
-    array<GlobalIndexType> send_positions{exec, num_entries};
-    array<GlobalIndexType> original_positions{exec, num_entries};
-    send_sizes.fill(zero<comm_index_type>());
-    exec->run(matrix::make_count_non_owning_entries(
-        input_data, row_partition, local_part, send_sizes, send_positions,
-        original_positions));
-
-    send_sizes.set_executor(exec->get_master());
-    array<comm_index_type> send_offsets{exec->get_master(), num_parts + 1};
-    array<comm_index_type> recv_sizes{exec->get_master(), num_parts};
-    array<comm_index_type> recv_offsets{exec->get_master(), num_parts + 1};
-
-    std::partial_sum(send_sizes.get_data(), send_sizes.get_data() + num_parts,
-                     send_offsets.get_data() + 1);
-    comm.all_to_all(exec, send_sizes.get_data(), 1, recv_sizes.get_data(), 1);
-    std::partial_sum(recv_sizes.get_data(), recv_sizes.get_data() + num_parts,
-                     recv_offsets.get_data() + 1);
-    send_offsets.get_data()[0] = 0;
-    recv_offsets.get_data()[0] = 0;
-
-    size_type n_send = send_offsets.get_data()[num_parts];
-    size_type n_recv = recv_offsets.get_data()[num_parts];
-    array<GlobalIndexType> send_row_idxs{exec, n_send};
-    array<GlobalIndexType> send_col_idxs{exec, n_send};
-    array<ValueType> send_values{exec, n_send};
-    array<GlobalIndexType> recv_row_idxs{exec, n_recv};
-    array<GlobalIndexType> recv_col_idxs{exec, n_recv};
-    array<ValueType> recv_values{exec, n_recv};
-    exec->run(matrix::make_fill_send_buffers(
-        input_data, row_partition, local_part, send_positions,
-        original_positions, send_row_idxs, send_col_idxs, send_values));
-
-    if (use_host_buffer) {
-        send_row_idxs.set_executor(exec->get_master());
-        send_col_idxs.set_executor(exec->get_master());
-        send_values.set_executor(exec->get_master());
-        recv_row_idxs.set_executor(exec->get_master());
-        recv_col_idxs.set_executor(exec->get_master());
-        recv_values.set_executor(exec->get_master());
-    }
-    auto row_req = comm.i_all_to_all_v(
-        use_host_buffer ? exec : exec->get_master(),
-        send_row_idxs.get_const_data(), send_sizes.get_data(),
-        send_offsets.get_data(), recv_row_idxs.get_data(),
-        recv_sizes.get_data(), recv_offsets.get_data());
-    auto col_req = comm.i_all_to_all_v(
-        use_host_buffer ? exec : exec->get_master(),
-        send_col_idxs.get_const_data(), send_sizes.get_data(),
-        send_offsets.get_data(), recv_col_idxs.get_data(),
-        recv_sizes.get_data(), recv_offsets.get_data());
-    auto val_req =
-        comm.i_all_to_all_v(use_host_buffer ? exec : exec->get_master(),
-                            send_values.get_const_data(), send_sizes.get_data(),
-                            send_offsets.get_data(), recv_values.get_data(),
-                            recv_sizes.get_data(), recv_offsets.get_data());
-
-    array<GlobalIndexType> all_row_idxs{exec, num_entries + n_recv};
-    array<GlobalIndexType> all_col_idxs{exec, num_entries + n_recv};
-    array<ValueType> all_values{exec, num_entries + n_recv};
-    exec->copy_from(exec, num_entries, input_data.get_const_row_idxs(),
-                    all_row_idxs.get_data());
-    exec->copy_from(exec, num_entries, input_data.get_const_values(),
-                    all_values.get_data());
-    exec->copy_from(exec, num_entries, input_data.get_const_col_idxs(),
-                    all_col_idxs.get_data());
-
-    row_req.wait();
-    col_req.wait();
-    val_req.wait();
-    if (use_host_buffer) {
-        recv_row_idxs.set_executor(exec);
-        recv_col_idxs.set_executor(exec);
-        recv_values.set_executor(exec);
-    }
-    exec->copy_from(exec, n_recv, recv_row_idxs.get_data(),
-                    all_row_idxs.get_data() + num_entries);
-    exec->copy_from(exec, n_recv, recv_col_idxs.get_data(),
-                    all_col_idxs.get_data() + num_entries);
-    exec->copy_from(exec, n_recv, recv_values.get_data(),
-                    all_values.get_data() + num_entries);
-    all_data = device_matrix_data<ValueType, GlobalIndexType>{
-        exec, global_dim, std::move(all_row_idxs), std::move(all_col_idxs),
-        std::move(all_values)};
-    all_data.sum_duplicates();
-}
-
-
 }  // namespace matrix
 
 
@@ -354,8 +246,7 @@ void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
     std::shared_ptr<const Partition<local_index_type, global_index_type>>
         row_partition,
     std::shared_ptr<const Partition<local_index_type, global_index_type>>
-        col_partition,
-    assembly_mode assembly_type)
+        col_partition)
 {
     const auto comm = this->get_communicator();
     GKO_ASSERT_EQ(data.get_size()[0], row_partition->get_size());
@@ -374,12 +265,6 @@ void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
     dim<2> global_dim{global_num_rows, global_num_cols};
     this->set_size(global_dim);
 
-    device_matrix_data<value_type, global_index_type> all_data{exec};
-    if (assembly_type == assembly_mode::communicate) {
-        matrix::communicate_non_owning(exec, comm, tmp_row_partition.get(),
-                                       tmp_col_partition.get(), data, all_data);
-    }
-
     // temporary storage for the output
     array<local_index_type> local_row_idxs{exec};
     array<local_index_type> local_col_idxs{exec};
@@ -393,8 +278,7 @@ void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
     // as well as the rows of the non-local block. The columns of the non-local
     // block are still in global indices.
     exec->run(matrix::make_separate_local_nonlocal(
-        assembly_type == assembly_mode::communicate ? all_data : data,
-        tmp_row_partition.get(), tmp_col_partition.get(), local_part,
+        data, tmp_row_partition.get(), tmp_col_partition.get(), local_part,
         local_row_idxs, local_col_idxs, local_values, non_local_row_idxs,
         global_non_local_col_idxs, non_local_values));
 
@@ -477,13 +361,25 @@ void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
     std::shared_ptr<const Partition<local_index_type, global_index_type>>
         row_partition,
     std::shared_ptr<const Partition<local_index_type, global_index_type>>
-        col_partition,
-    assembly_mode assembly_type)
+        col_partition)
 {
     return this->read_distributed(
         device_matrix_data<value_type, global_index_type>::create_from_host(
             this->get_executor(), data),
-        row_partition, col_partition, assembly_type);
+        row_partition, col_partition);
+}
+
+
+template <typename ValueType, typename LocalIndexType, typename GlobalIndexType>
+void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
+    const matrix_data<ValueType, global_index_type>& data,
+    std::shared_ptr<const Partition<local_index_type, global_index_type>>
+        partition)
+{
+    return this->read_distributed(
+        device_matrix_data<value_type, global_index_type>::create_from_host(
+            this->get_executor(), data),
+        partition, partition);
 }
 
 
@@ -494,10 +390,28 @@ void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
         partition,
     assembly_mode assembly_type)
 {
-    return this->read_distributed(
+    if (assembly_type == assembly_mode::local_only) {
+        return this->read_distributed(
+            device_matrix_data<value_type, global_index_type>::create_from_host(
+                this->get_executor(), data),
+            partition, partition);
+    }
+    auto all_data = assemble<ValueType, LocalIndexType, GlobalIndexType>(
+        this->get_communicator(),
         device_matrix_data<value_type, global_index_type>::create_from_host(
             this->get_executor(), data),
-        partition, partition, assembly_type);
+        partition);
+    return this->read_distributed(all_data, partition, partition);
+}
+
+
+template <typename ValueType, typename LocalIndexType, typename GlobalIndexType>
+void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
+    const device_matrix_data<ValueType, GlobalIndexType>& data,
+    std::shared_ptr<const Partition<local_index_type, global_index_type>>
+        partition)
+{
+    return this->read_distributed(data, partition, partition);
 }
 
 
@@ -508,7 +422,12 @@ void Matrix<ValueType, LocalIndexType, GlobalIndexType>::read_distributed(
         partition,
     assembly_mode assembly_type)
 {
-    return this->read_distributed(data, partition, partition, assembly_type);
+    if (assembly_type == assembly_mode::local_only) {
+        return this->read_distributed(data, partition, partition);
+    }
+    auto all_data = assemble<ValueType, LocalIndexType, GlobalIndexType>(
+        this->get_communicator(), data, partition);
+    return this->read_distributed(all_data, partition, partition);
 }
