+++ dpcpp/distributed/matrix_kernels.dp.cpp
@@ -14,36 +14,6 @@ namespace distributed_matrix {
 
 
 template <typename ValueType, typename LocalIndexType, typename GlobalIndexType>
-void count_non_owning_entries(
-    std::shared_ptr<const DefaultExecutor> exec,
-    const device_matrix_data<ValueType, GlobalIndexType>& input,
-    const experimental::distributed::Partition<LocalIndexType, GlobalIndexType>*
-        row_partition,
-    comm_index_type local_part, array<comm_index_type>& send_count,
-    array<GlobalIndexType>& send_positions,
-    array<GlobalIndexType>& original_positions) GKO_NOT_IMPLEMENTED;
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_LOCAL_GLOBAL_INDEX_TYPE(
-    GKO_DECLARE_COUNT_NON_OWNING_ENTRIES);
-
-
-template <typename ValueType, typename LocalIndexType, typename GlobalIndexType>
-void fill_send_buffers(
-    std::shared_ptr<const DefaultExecutor> exec,
-    const device_matrix_data<ValueType, GlobalIndexType>& input,
-    const experimental::distributed::Partition<LocalIndexType, GlobalIndexType>*
-        row_partition,
-    comm_index_type local_part, const array<GlobalIndexType>& send_positions,
-    const array<GlobalIndexType>& original_positions,
-    array<GlobalIndexType>& send_row_idxs,
-    array<GlobalIndexType>& send_col_idxs,
-    array<ValueType>& send_values) GKO_NOT_IMPLEMENTED;
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_LOCAL_GLOBAL_INDEX_TYPE(
-    GKO_DECLARE_FILL_SEND_BUFFERS);
-
-
-template <typename ValueType, typename LocalIndexType, typename GlobalIndexType>
 void separate_local_nonlocal(
     std::shared_ptr<const DefaultExecutor> exec,
     const device_matrix_data<ValueType, GlobalIndexType>& input,
