+++ include/ginkgo/core/distributed/matrix.hpp
@@ -303,6 +303,27 @@ public:
      *
      * @param data  The device_matrix_data structure.
      * @param partition  The global row and column partition.
+     *
+     * @return the index_map induced by the partitions and the matrix structure
+     */
+    void read_distributed(
+        const device_matrix_data<value_type, global_index_type>& data,
+        std::shared_ptr<const Partition<local_index_type, global_index_type>>
+            partition);
+
+    /**
+     * Reads a square matrix from the device_matrix_data structure and a global
+     * partition.
+     *
+     * The global size of the final matrix is inferred from the size of the
+     * partition. Both the number of rows and columns of the device_matrix_data
+     * are ignored.
+     *
+     * @note The matrix data can contain entries for rows other than those owned
+     *        by the process. Entries for those rows are discarded.
+     *
+     * @param data  The device_matrix_data structure.
+     * @param partition  The global row and column partition.
      * @param assembly_mode  The mode of assembly.
      *
      * @return the index_map induced by the partitions and the matrix structure
@@ -311,7 +332,21 @@ public:
         const device_matrix_data<value_type, global_index_type>& data,
         std::shared_ptr<const Partition<local_index_type, global_index_type>>
             partition,
-        assembly_mode assembly_type = assembly_mode::local_only);
+        assembly_mode assembly_type);
+
+    /**
+     * Reads a square matrix from the matrix_data structure and a global
+     * partition.
+     *
+     * @see read_distributed
+     *
+     * @note For efficiency it is advised to use the device_matrix_data
+     * overload.
+     */
+    void read_distributed(
+        const matrix_data<value_type, global_index_type>& data,
+        std::shared_ptr<const Partition<local_index_type, global_index_type>>
+            partition);
 
     /**
      * Reads a square matrix from the matrix_data structure and a global
@@ -326,7 +361,7 @@ public:
         const matrix_data<value_type, global_index_type>& data,
         std::shared_ptr<const Partition<local_index_type, global_index_type>>
             partition,
-        assembly_mode assembly_type = assembly_mode::local_only);
+        assembly_mode assembly_type);
 
     /**
      * Reads a matrix from the device_matrix_data structure, a global row
@@ -351,8 +386,7 @@ public:
         std::shared_ptr<const Partition<local_index_type, global_index_type>>
             row_partition,
         std::shared_ptr<const Partition<local_index_type, global_index_type>>
-            col_partition,
-        assembly_mode assembly_type = assembly_mode::local_only);
+            col_partition);
 
     /**
      * Reads a matrix from the matrix_data structure, a global row partition,
@@ -368,8 +402,7 @@ public:
         std::shared_ptr<const Partition<local_index_type, global_index_type>>
             row_partition,
         std::shared_ptr<const Partition<local_index_type, global_index_type>>
-            col_partition,
-        assembly_mode assembly_type = assembly_mode::local_only);
+            col_partition);
 
     /**
      * Get read access to the stored local matrix.
