+++ reference/test/distributed/CMakeLists.txt
@@ -1,3 +1,4 @@
+ginkgo_create_test(assembly_helpers_kernels)
 ginkgo_create_test(index_map_kernels)
 ginkgo_create_test(matrix_kernels)
 ginkgo_create_test(partition_helpers_kernels)
