+++ test/distributed/CMakeLists.txt
@@ -1,3 +1,4 @@
+ginkgo_create_common_test(assembly_helpers_kernels DISABLE_EXECUTORS dpcpp)
 ginkgo_create_common_test(index_map_kernels DISABLE_EXECUTORS dpcpp)
 ginkgo_create_common_test(matrix_kernels DISABLE_EXECUTORS dpcpp)
 ginkgo_create_common_test(partition_kernels DISABLE_EXECUTORS dpcpp)
